feat: viewmodel의 service 로직 분리
